[#120] Add CredentialsLoader for MailJet API keys
This credentials loader will determine if the application is running locally or on Google App Engine.

If running locally the credentials will be loaded from the environment variables "MJ_APIKEY_PRIVATE" and "MJ_APIKEY_PUBLIC".
If running on Google App Engine, the credentials will be loaded from the credential store from within the project. The user running the app needs to have read permission on the respective secrets.