[#128, #129] Add image upload capabilities
A post to /plants/<plantId>/image with 1 MultipartFile that is sent with the parameter **"image"** will upload this file to Google Cloud Storage and return the URL to the file.
The URL will additionally be added to the field "plantImageUrl" for the plant with <plantId>. This attribute is returned with every plant object.

 The submitted file MUST contain an image with ".jpg", ".jpeg", ".png" ending and is smaller than 2MB.

 Closes 128, Closes 129